Implementada la funcionalidad para añadir tareas y ver los detalles de las mismas. Nota: FALTA COMPROBAR LOS RESULTADOS Y MEJORAR LA NAVEGACIÓN.

+++ DSDM/Task.h
@@ -13,11 +13,11 @@
 @property (weak, nonatomic) NSString* name;
 @property (weak, nonatomic) NSDate* date;
 @property (weak, nonatomic) NSString* note;
-@property (nonatomic) NSInteger priority;
+@property (nonatomic) float priority;
 @property (nonatomic) BOOL alreadyDone;
 
 - (id)init;
-- (id)initWithName:(NSString*)name date:(NSDate*)date note:(NSString*)note priority:(NSInteger)priority;
+- (id)initWithName:(NSString*)name date:(NSDate*)date note:(NSString*)note priority:(float)priority;
 
 
 @end
